v6.3

+++ plugins/massiveAdmin/inc/newOptionsMassive.inc.php
@@ -1,3 +1,11 @@
+<?php
+// ADDED: Generate CSRF tokens
+$copy_nonce = MassiveAdminClass::generate_nonce('copy_rename');
+$delete_nonce = MassiveAdminClass::generate_nonce('delete_file');
+$upload_nonce = MassiveAdminClass::generate_nonce('upload_file');
+$save_nonce = MassiveAdminClass::generate_nonce('save_rename');
+?>
+
 <style>
 	@import url('<?php global $SITEURL;
 	echo $SITEURL; ?>plugins/massiveAdmin/css/newOptionsMassive.css');
@@ -6,6 +14,9 @@
 <div class="rename-fog hide-fog">
 	<div class="form-rename">
 		<form class="form-form-rename" action="#" method="post">
+			<!-- ADDED: CSRF token field -->
+			<input type="hidden" name="nonce" value="<?php echo htmlspecialchars($copy_nonce); ?>">
+			
 			<input type="text" name="rename-massive-hide" style="display:none">
 			<input type="text" name="rename-massive">
 			<input type="submit" name="save-rename-massive" class="submit" value="<?php echo i18n_r("massiveAdmin/RENAMEFILE"); ?>">
@@ -19,20 +30,37 @@
 <script src="<?php echo $SITEURL; ?>plugins/massiveAdmin/js/newOptionsMassive.js?v=6"></script>
 
 <?php
-
+// MODIFIED: Added CSRF checks before processing
 if (isset($_POST['deleteFileList'])) {
+	if (!isset($_POST['nonce']) || !MassiveAdminClass::verify_nonce($_POST['nonce'], 'delete_file')) {
+		die('CSRF validation failed');
+	}
 	global $MA;
 	$MA->deleteFileList();
 };
 
 if (isset($_POST['save-rename-massive'])) {
+	if (!isset($_POST['nonce']) || !MassiveAdminClass::verify_nonce($_POST['nonce'], 'save_rename')) {
+		die('CSRF validation failed');
+	}
 	global $MA;
 	$MA->saveRename();
 }
 
 if (isset($_POST['copy-rename-massive'])) {
+	if (!isset($_POST['nonce']) || !MassiveAdminClass::verify_nonce($_POST['nonce'], 'copy_rename')) {
+		die('CSRF validation failed');
+	}
 	global $MA;
 	$MA->copyRename();
 };
 
+// ADDED: Handle uploads with CSRF check
+if (isset($_POST['massiveUpload'])) {
+	if (!isset($_POST['nonce']) || !MassiveAdminClass::verify_nonce($_POST['nonce'], 'upload_file')) {
+		die('CSRF validation failed');
+	}
+	global $MA;
+	$MA->massiveUpload();
+}
 ?>
